pyproject.toml
[tool.poetry]
name = "cdp-merged"
version = "0.1.0"
description = "AI-powered Customer Data Platform with natural language query capabilities"
authors = ["CDP Team"]
readme = "README.md"
packages = [{include = "src"}]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
# Core frameworks
langgraph = ">=1.0.8,<2.0.0"
langchain-openai = ">=1.1.9,<2.0.0"
langchain-ollama = ">=1.0.1,<2.0.0"
chainlit = ">=2.9.6,<3.0.0"
# HTTP clients
httpx = ">=0.26.0,<0.29.0"
# Data validation
pydantic = ">=2.11,<3"
pydantic-settings = ">=2.1.0,<3.0.0"
# Elasticsearch
elasticsearch = ">=8.11.0,<9.0.0"
# PostgreSQL
asyncpg = ">=0.30.0,<1.0.0"
# Resilience & retries
tenacity = ">=8.2.0,<9.0.0"
# Structured logging
structlog = ">=24.1.0,<25.0.0"
# Metrics
prometheus-client = ">=0.20.0,<1.0.0"
azure-identity = ">=1.17.0,<2.0.0"
azure-keyvault-secrets = ">=4.8.0,<5.0.0"
# HTML parsing for enrichment
beautifulsoup4 = ">=4.12.0,<5.0.0"
lxml = ">=6.0"
# HTTP client
requests = "^2.32.0"
aiohttp = "^3.13.3"
redis = "^7.2.1"
aiosqlite = ">=0.21,<0.22"
langgraph-checkpoint-sqlite = ">=3.0.3,<4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0,<9.0.0"
pytest-asyncio = ">=0.23.0,<0.25.0"
pytest-mock = ">=3.12.0,<4.0.0"
pytest-cov = ">=5.0.0,<6.0.0"
coverage = {extras = ["toml"], version = ">=7.4.0,<8.0.0"}
ruff = ">=0.4.0,<1.0.0"
mypy = ">=1.8.0,<2.0.0"
pre-commit = ">=3.6.0,<4.0.0"
factory-boy = ">=3.3.0,<4.0.0"
sqlalchemy = ">=2.0.0,<3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ─── Ruff ────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "W",   # pycodestyle warnings
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in default arguments
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─── MyPy ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true
exclude = ["tests/", "scripts/"]

# ─── Pytest ───────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
markers = [
    "unit: pure unit tests with no external dependencies",
    "integration: tests requiring running external services",
    "e2e: full end-to-end tests",
    "slow: tests that take a long time",
]
addopts = "--tb=short -q"
# TODO: Remove when SQLAlchemy is added to dependencies and repository layer is active
# These tests require sqlalchemy which is not yet in the project's dependencies
norecursedirs = ["tests/unit/repository"]

# ─── Coverage ─────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
